+++ CHANGELOG.md
@@ -1,13 +1,22 @@
 # CHANGELOG
 
+## 2.8.0
+#### New Functionality
+
+- [#621](https://github.com/marklogic/node-client-api/issues/621)- Data Movement in Node.js API - Ingesting Documents using writeAll api.
+
+#### Improvements and Bug Fixes
+- [#270](https://github.com/marklogic/node-client-api/issues/270) - Calling methods on resources with arg list throws error
+- [#601](https://github.com/marklogic/node-client-api/issues/601) - Regenerate Optic expression functions
+
+
 ## 2.7.0
 #### New Functionality
 
 - [#543](https://github.com/marklogic/node-client-api/issues/543) - Node-client-api now prefers ipv4 over ipv6
 - [#562](https://github.com/marklogic/node-client-api/issues/562) - CtsQueryBuilder is available
 - [#567](https://github.com/marklogic/node-client-api/issues/567) - Dynamic interface to Data Service endpoints
-- [#569](https://github.com/marklogic/node-client-api/issues/569) - Optic enhancements in 10.0-7
-
+- [#569](https://github.com/marklogic/node-client-api/issues/569) - Optic enhancements in 10.0-7 - Server release 10.0-7 or above required
 
 #### Improvements and Bug Fixes
 - [#485](https://github.com/marklogic/node-client-api/issues/485) - Throw error when complexValues is set with a chunked stream